Search fix (#27)
* Fixed bug in search

It was possible that none of the moves in the AB search move loop would run due to heavy pruning. This in turn resulted in the lower score bound being too pessimistic after the loop. The fix guarantees that the first move will always get tried.

Elo performance of the change was measured as neutral. The fix removed obscure behavior noted in some positions.

* Updated version number