CNN Pawn Eval (#31)
Winter eval now consists of a hybrid neural network. Pawn structure evaluation is for the most part taken care of by a convolutional neural network. The rest of the evaluation function relies on a neural network taking handcrafted features as input.

There is now a pawn hash to speedup evaluation.

Likely a negative patch on lower hardware and faster TC, but scaling fantastically. Likely neutral under TCEC conditions. Decided to merge as I see longter, potential and it is an interesting experiment.

* Added pawn hash

* Added support to add pawns and kings as raw input for a potential CNN

* Updating net_evaluation for CNN inputs

* Fully functioning hybrid network

* Saving previous work as it has been a long time

* Testing new CNN architecture

* Updated net

* Makefile now uses -O3 instead of -Ofast and has config for target older crosscompile

* Updated version number