Incbin net (#47)
* Added support for fully connected incbin based network
* Updated net_evaluation to properly handle block size
* Fully switched to new eval code
* Fixed order of makefile arguments
* Updated Winter net and version.

Winter net is now a simple fully connected network. Incbin is used to load the weights in binary form. Overall the new net is significantly weaker at regular chess but slightly better at DFRC, as it was trained with DFRC games. Hopefully this will be a long-term gain at the cost of a fairly temporary strength regression.

Bench 1303742